Allow to set priority via text or picker when creating reminder (#207)
* Allow setting priority

Makes use of new dependency WrappingHStack to ensure that date, time and priority don't get squooshed.

Rather than use a picker (which didn't really look too stylish!) the user clicks to cycle None -> Low -> Medium -> High; the icon updates too.

* Add PriorityParser (!)

Entirely functional! That said, if you type too many !s, or type a ! and then delete it, it lingers with a priority, which is a little annoying.

* Removing usage of WrappingHStack dependency

* Simplifying logic using EKReminderPriority extensions

* Extracting NewReminderInfoOptionsView to a separate view

* Fixing project.pbxproj

* Fix scenarios for PriorityParser and simplify logic

* Changing parameter order according to contexts

* Fixing replacing in title of PriorityParser detection

---------

Co-authored-by: Rafael Damasceno <damscn@gmail.com>